fix(prices): stop pulling server-only env into the PricesView client bundle
PricesView is a client component but imported formatLedgerInstant from
utils/formatDate, whose configured-locale default transitively imports
'server-only' via @/lib/env — failing the Next 16 Turbopack build with
"'server-only' cannot be imported from a Client Component module". CI
didn't catch it (lint + test only, no next build).

Move the pure formatLedger* helpers into the client-safe utils/formatDateCore;
formatDate re-exports them for server callers, and PricesView imports them from
the core module. Build now compiles with no server-only-in-client error.
